# Conflicts:
#	src/main/java/com/SecurityConfig.java
#	src/main/java/com/controllers/ClaimConfirmController.java
#	src/main/java/com/controllers/IndexController.java
#	src/main/java/com/repository/ProductRepository.java
#	src/main/java/com/serviceImpl/BankBillCreateImpl.java
#	src/main/resources/templates/edit-profile.html
#	src/main/resources/templates/index.html

+++ src/main/java/com/entity/User.java
@@ -48,6 +48,7 @@ public class User {
 
     public User() {
     }
+
     public User(String name, String email, String password) {
         this.name = name;
         this.email = email;
@@ -58,7 +59,7 @@ public class User {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
